_makingInvest_data/makinginvest_server_data_python/requirements.txt
aiocron==1.8
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
black==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
ccxt==2.2.46
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
croniter==1.3.8
cryptography==38.0.4
cycler==0.11.0
dnspython==2.2.1
fastapi==0.88.0
fastapi-socketio==0.0.9
firebase-admin==6.0.1
fonttools==4.38.0
frozenlist==1.3.3
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httplib2==0.21.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
motor==3.1.1
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pandas-ta==0.3.14b0
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
socketIO-client==0.7.2
starlette==0.22.0
timeago==1.0.16
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
websocket-client==1.4.2
websockets==10.4
yarl==1.8.1
